+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/rendering/TextStyle.kt
@@ -1,29 +1,65 @@
 package com.github.ajalt.mordant.rendering
 
 import com.github.ajalt.colormath.Color
+import com.github.ajalt.colormath.model.Ansi16
+import com.github.ajalt.mordant.internal.*
+import com.github.ajalt.mordant.internal.AnsiCodes
 import com.github.ajalt.mordant.internal.DEFAULT_STYLE
+import com.github.ajalt.mordant.internal.HYPERLINK_RESET
 import com.github.ajalt.mordant.internal.generateHyperlinkId
 import com.github.ajalt.mordant.internal.invokeStyle
 
+/**
+ * Styles that can be applied to text by terminals that support it.
+ *
+ * You can combine styles with [on] or [plus]
+ */
 interface TextStyle {
     val color: Color?
     val bgColor: Color?
-    val bold: Boolean
-    val italic: Boolean
-    val underline: Boolean
-    val dim: Boolean
-    val inverse: Boolean
-    val strikethrough: Boolean
+    val bold: Boolean?
+    val italic: Boolean?
+    val underline: Boolean?
+    val dim: Boolean?
+    val inverse: Boolean?
+    val strikethrough: Boolean?
     val hyperlink: String?
     val hyperlinkId: String?
 
     val bg: TextStyle
 
+    /**
+     * Create a style with the foreground color of [bg] as the background color of the new style.
+     *
+     * All other attributes will be copied from this style.
+     *
+     * ### Example
+     * ```
+     * (red on blue) == TextStyle(color=red, bgColor=blue)
+     *
+     * val style1 = TextStyle(red, bold=true)
+     * val style2 = TextStyle(blue, italic=true)
+     * (style1 on style2) == TextStyle(red, blue, bold=true)
+     * ```
+     */
     infix fun on(bg: TextStyle): TextStyle
 
     operator fun invoke(text: String): String = invokeStyle(text)
 
-    operator fun plus(other: TextStyles): TextStyle = this + other.style
+
+    /**
+     * Return a new style that combines this style with [other].
+     *
+     * And attributes that are set on both this and [other] will take the value from [other]. Any
+     * `null` values on [other] will keep the value from this style.
+     *
+     * ### Example
+     * ```
+     * val style1 = TextStyle(red, blue, bold=true, italic=true)
+     * val style2 = TextStyle(green, italic=false, dim=true)
+     * (style1 + style2) == TextStyle(green, blue, bold=true, italic=false, dim=true)
+     * ```
+     */
     operator fun plus(other: TextStyle): TextStyle {
         return when {
             this === DEFAULT_STYLE -> other
@@ -31,29 +67,30 @@ interface TextStyle {
             else -> TxtStyle(
                 color = other.color ?: color,
                 bgColor = other.bgColor ?: bgColor,
-                bold = other.bold || bold,
-                italic = other.italic || italic,
-                underline = other.underline || underline,
-                dim = other.dim || dim,
-                inverse = other.inverse || inverse,
-                strikethrough = other.strikethrough || strikethrough,
+                bold = other.bold ?: bold,
+                italic = other.italic ?: italic,
+                underline = other.underline ?: underline,
+                dim = other.dim ?: dim,
+                inverse = other.inverse ?: inverse,
+                strikethrough = other.strikethrough ?: strikethrough,
                 hyperlink = other.hyperlink ?: hyperlink,
                 hyperlinkId = other.hyperlinkId ?: hyperlinkId
             )
         }
     }
+    operator fun plus(other: TextStyles): TextStyle = this + other.style
 }
 
 @Suppress("FunctionName")
 fun TextStyle(
     color: Color? = null,
     bgColor: Color? = null,
-    bold: Boolean = false,
-    italic: Boolean = false,
-    underline: Boolean = false,
-    dim: Boolean = false,
-    inverse: Boolean = false,
-    strikethrough: Boolean = false,
+    bold: Boolean? = null,
+    italic: Boolean? = null,
+    underline: Boolean? = null,
+    dim: Boolean? = null,
+    inverse: Boolean? = null,
+    strikethrough: Boolean? = null,
     hyperlink: String? = null,
 ): TextStyle = TxtStyle(
     color = color,
@@ -71,12 +108,12 @@ fun TextStyle(
 internal data class TxtStyle(
     override val color: Color?,
     override val bgColor: Color?,
-    override val bold: Boolean,
-    override val italic: Boolean,
-    override val underline: Boolean,
-    override val dim: Boolean,
-    override val inverse: Boolean,
-    override val strikethrough: Boolean,
+    override val bold: Boolean?,
+    override val italic: Boolean?,
+    override val underline: Boolean?,
+    override val dim: Boolean?,
+    override val inverse: Boolean?,
+    override val strikethrough: Boolean?,
     override val hyperlink: String?,
     override val hyperlinkId: String?,
 ) : TextStyle {