+++ mordant/src/commonTest/kotlin/com/github/ajalt/mordant/table/LinearLayoutTest.kt
@@ -5,7 +5,7 @@ import com.github.ajalt.mordant.rendering.TextAlign.CENTER
 import com.github.ajalt.mordant.rendering.TextColors.blue
 import com.github.ajalt.mordant.rendering.TextColors.red
 import com.github.ajalt.mordant.rendering.VerticalAlign.MIDDLE
-import com.github.ajalt.mordant.rendering.Whitespace.PRE
+import com.github.ajalt.mordant.rendering.Whitespace.*
 import com.github.ajalt.mordant.table.Borders.*
 import com.github.ajalt.mordant.terminal.Terminal
 import com.github.ajalt.mordant.test.RenderingTest
@@ -20,11 +20,16 @@ class LinearLayoutTest : RenderingTest() {
     fun horizontalLayoutEmpty() = checkRender(
         horizontalLayout {}, "", width = 10
     )
+
     @Test
     fun horizontalLayout() = checkRender(
         horizontalLayout {
             column(1) { width = ColumnWidth.Expand() }
-            val bar = ProgressBar(indeterminate = true, showPulse = false, indeterminateStyle = TextStyle())
+            val bar = ProgressBar(
+                indeterminate = true,
+                showPulse = false,
+                indeterminateStyle = TextStyle()
+            )
             cells(1, bar, 2)
         }, """
         ░1 ━━━━━━ 2░
@@ -36,7 +41,11 @@ class LinearLayoutTest : RenderingTest() {
         horizontalLayout {
             spacing = 2
             column(1) { width = ColumnWidth.Expand() }
-            val bar = ProgressBar(indeterminate = true, showPulse = false, indeterminateStyle = TextStyle())
+            val bar = ProgressBar(
+                indeterminate = true,
+                showPulse = false,
+                indeterminateStyle = TextStyle()
+            )
             cells(1, bar, 2)
         }, """
         ░1  ━━━━  2░
@@ -49,18 +58,51 @@ class LinearLayoutTest : RenderingTest() {
         ░
         """
     )
+
     @Test
-    fun verticalLayout() = checkRender(
+    fun verticalLayoutSimple() = checkRender(
+        verticalLayout {
+            cell("1111")
+            cells("2", "3")
+            cellsFrom(listOf(Text("4 4")))
+        }, """
+        ░1111░
+        ░2░
+        ░3░
+        ░4 4░
+        """
+    )
+
+    @Test
+    fun verticalLayoutCellAlign() = checkRender(
         verticalLayout {
             cell("1111")
             cells("2", "3") { align = TextAlign.RIGHT }
-            cellsFrom(listOf(Text("4 4")))
+            cell(Text("4", align = TextAlign.LEFT))
+            cellsFrom(listOf(Text("5 5")))
         }, """
         ░1111░
         ░   2░
         ░   3░
-        ░4 4░
+        ░4   ░
+        ░5 5░
         """
+    )
+
+    @Test
+    fun verticalLayoutCellAlignExpand() = checkRender(
+        verticalLayout {
+            width = ColumnWidth.Expand()
+            align = TextAlign.RIGHT
+            cell("1")
+            cell("2") { align = TextAlign.RIGHT }
+            cell(Text("3", align = TextAlign.LEFT))
+        }, """
+        ░     1░
+        ░     2░
+        ░3     ░
+        """,
+        width = 6
     )
 
     @Test
@@ -78,6 +120,7 @@ class LinearLayoutTest : RenderingTest() {
         ░3░
         """
     )
+
     @Test
     fun verticalLayoutLeftAlign() = checkRender(
         verticalLayout {
@@ -94,6 +137,18 @@ class LinearLayoutTest : RenderingTest() {
         ░33░
         """
     )
+
+    @Test
+    fun verticalLayoutFixedTruncation() = checkRender(
+        verticalLayout {
+            whitespace = Whitespace.NOWRAP
+            cell("1111 222")
+            width = ColumnWidth.Fixed(6)
+        }, """
+        ░1111 2░
+        """
+    )
+
 
     @Test
     @JsName("nesting_horizontalLayouts_in_verticalLayouts")